Finished telemarkers
Also got rid of annoying console spam. I just commented them out so if
you need them theyre still there. Just try to keep it to a minimum
because its a pain in the ass to sift thru your code to get rid of it,
especially when im trying to debug too.